requirements.txt
Django>=5.2.2
requests>=2.32.3
python-dotenv>=1.1.0
kubernetes>=32.0.1
prometheus-client>=0.22.1
python-jenkins>=1.8.0
gunicorn>=21.2.0
whitenoise>=6.6.0
django-cors-headers>=4.3.1
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
python-dateutil>=2.8.2
pytz>=2024.1 